+++ forge/changelog.md
@@ -1,2 +1,1 @@
-- Updated Japanese translation
-- Fixed MTU size not considered when reading voice chat packets
+- Added volume config accessor to the API